Handle VCs that hash to 0 (#123) (#124)
* Handle VCs that hash to 0

Because of a mistake (perhaps because phash/1 hashed to a range starting at 1, and phash2/1 now starts at 0) - special meaning was attributed to a hash of 0, which would means that updating hashtrees would not be handled correctly if the actual clock had a hash collision with 0.

Special meaning now uses `none` not 0 - and 0 is handled as with any other hash value.

* Add support for CH, OH of none, none

Used in rehash in riak_kv_vnode - expected to be a neutral change i.e. CH, CH

* Update comment

* Clarify comments/types

* Update GHA and workflow

* Update erlang.yml