Add ties to the set of selections to use (#126)
Will randomly select from the top MaxResults set of ID clusters, and any tied with the sparsest set within the MaxResults.

This means when there is a contiguous set of IDs (e.g. when doing root_confirm with a fully broken root), any set of IDs will be chosen, not just sets within the first MaxResults.  This ensures a better spread of selections across the tree, so that the mismatched_segments will more accurately reflect the overall state of the tree